rasa-chatbot/requirements.txt
rasa==3.6.15
rasa-sdk==3.6.2
python-socketio==5.10.0
python-engineio==4.7.1
requests==2.31.0
